Extract Abstract Repository
